fix(admin): 修复文件名清理逻辑并更新用户查询接口
- 将文件名清理正则表达式从仅去除空格改为过滤所有特殊字符，保留中英文数字
- 在歌曲上传和更新功能中统一应用新的正则清理逻辑
- 修改用户查询接口参数类型从 Integer 改为 Long 以支持更大范围ID查询

+++ src/main/java/com/music/mapper/UserMapper.java
@@ -41,5 +41,5 @@ public interface UserMapper {
 
     // 8. 顺手加个按 ID 查询，用来返回最新数据
     @org.apache.ibatis.annotations.Select("SELECT * FROM user_info WHERE id = #{id}")
-    com.music.entity.User selectById(Integer id);
+    com.music.entity.User selectById(Long id);
 }
